Ligera reestructuracion a TODO
se hicieron mejoras a DibujadoJuego (solo se cambio una línea).

Se hicieron ligeros cambios a la estructura del nivel para generar el GestorJuego y GestorEditor.

+++ core/src/main/java/io/github/rubfv/uchitai/NotaSostenida.java
@@ -4,14 +4,14 @@ package io.github.rubfv.uchitai;
 
 //Aquellas teclas que duran una x cantidad de milisegundos presionadas
 public class NotaSostenida extends Nota{
-    private long tiempoFinal;       //Momento cuando se debe soltar la tecla
+    private float tiempoFinal;       //Momento cuando se debe soltar la tecla
     private boolean inicioAcertado;
 
 
     // Todo CONSTRUCTOR (RUBÉN)
 
 
-    public long getTiempoFinal() {
+    public float getTiempoFinal() {
         return tiempoFinal;
     }
